requirements.txt
aiogram==2.10.1
aiohttp==3.6.3
aiohttp-healthcheck==1.3.1
aiohttp-jinja2==1.2.0
aiojobs==0.2.2
aiomysql==0.0.20
aiopg==1.0.0
aioredis==1.3.1
amqp==5.0.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
gunicorn==20.0.4
hiredis==1.1.0
idna==2.10
importlib-metadata==2.0.0
Jinja2==2.11.2
kombu==5.0.2
loguru==0.5.3
MarkupSafe==1.1.1
multidict==4.7.6
peewee==3.13.3
peewee-async==0.7.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==0.9.2
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
telegraph==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
yarl==1.5.1
zipp==3.4.0
